% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunShinyWizzard.R
\name{RunShinyWizzard}
\alias{RunShinyWizzard}
\title{Run ShinyWizzard App}
\usage{
RunShinyWizzard(dir = NULL)
}
\arguments{
\item{dir}{directory to tabs, when NULL run demo}
}
\description{
Run ShinyWizzard App
}
