% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyWizzardIncludeCSSAndTheme.R
\name{ShinyWizzardIncludeCSSAndTheme}
\alias{ShinyWizzardIncludeCSSAndTheme}
\title{ShinyWizzardIncludeCSSAndTheme}
\usage{
ShinyWizzardIncludeCSSAndTheme(ShinyWizzardTheme)
}
\arguments{
\item{ShinyWizzardTheme}{ShinyWizzard theme name returned by config$ShinyWizzardTheme}
}
\description{
ShinyWizzardIncludeCSSAndTheme
}
