% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyWizzardTheme.R
\name{ShinyWizzardTheme}
\alias{ShinyWizzardTheme}
\title{ShinyWizzard themes}
\usage{
ShinyWizzardTheme(name = "")
}
\arguments{
\item{name}{theme name to be applied. No argument returns all available themes.}
}
\description{
Load one of available themes for ShinyWizard project
}
