% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyWizzardTheme.R
\name{ShinyWizzardThemeSelector}
\alias{ShinyWizzardThemeSelector}
\title{ShinyWizzard themes selector}
\usage{
ShinyWizzardThemeSelector()
}
\description{
Shiny selector for themes. It returns a shiny ui code.
}
