% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunShinyWizzard.R
\name{RunShinyWizard}
\alias{RunShinyWizard}
\title{Run ShinyWizard app}
\usage{
RunShinyWizard(loc = "")
}
\arguments{
\item{loc}{directory to external tabs, when NULL run demo}
}
\description{
Run ShinyWizard app
}
