% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyWizzardIncludeCSSAndTheme.R
\name{ShinyWizzardIncludeCSSAndTheme}
\alias{ShinyWizzardIncludeCSSAndTheme}
\title{Run ShinyWizard app}
\usage{
ShinyWizzardIncludeCSSAndTheme()
}
\arguments{
\item{name}{the theme name to be applied. None argument provided returns all available themes.}
}
\description{
Run ShinyWizard app
}
