% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyWizzardTheme.R
\name{ShinyWizzardTheme}
\alias{ShinyWizzardTheme}
\title{ShinyWizzardTheme}
\usage{
ShinyWizzardTheme(name = "")
}
\arguments{
\item{name}{the theme name to be applied. None argument provided returns all available themes.}
}
\description{
Return dir / show available ShinyWizard templates
}
